redact numeric sequences and uuids from plausible analytics (#5240)
