+++ src/Components/Assets/AssetsList.tsx
@@ -12,9 +12,7 @@ import { getAsset } from "../../Redux/actions";
 import { useState, useCallback, useEffect } from "react";
 import { navigate } from "raviger";
 import loadable from "@loadable/component";
-import { make as SlideOver } from "../Common/SlideOver.gen";
 import AssetFilter from "./AssetFilter";
-import AdvancedFilterButton from "../Common/AdvancedFilterButton";
 import { parseQueryParams } from "../../Utils/primitives";
 import Chip from "../../CAREUI/display/Chip";
 import SearchInput from "../Form/SearchInput";
@@ -29,6 +27,7 @@ import ExportMenu from "../Common/Export";
 import CountBlock from "../../CAREUI/display/Count";
 import AssetImportModal from "./AssetImportModal";
 import Page from "../Common/components/Page";
+import { AdvancedFilterButton } from "../../CAREUI/interactive/FiltersSlideover";
 
 const Loading = loadable(() => import("../Common/Loading"));
 
@@ -341,7 +340,7 @@ const AssetsList = () => {
           <div className="flex flex-col md:flex-row gap-2 w-full lg:w-auto">
             <div className="w-full">
               <AdvancedFilterButton
-                setShowFilters={() => advancedFilter.setShow(true)}
+                onClick={() => advancedFilter.setShow(true)}
               />
             </div>
             <ButtonV2
@@ -368,13 +367,7 @@ const AssetsList = () => {
           </div>
         </div>
       </div>
-      <div>
-        <SlideOver {...advancedFilter}>
-          <div className="bg-white min-h-screen p-4">
-            <AssetFilter {...advancedFilter} />
-          </div>
-        </SlideOver>
-      </div>
+      <AssetFilter {...advancedFilter} />
       {isLoading ? (
         <Loading />
       ) : (
