+++ src/Components/Common/prescription-builder/PrescriptionMultiselect.tsx
@@ -82,9 +82,10 @@ export function PrescriptionMultiDropdown(props: {
                 type="button"
                 key={i}
                 className={classNames(
-                  "w-full block px-4 py-2 text-sm leading-5 text-left text-gray-700 hover:bg-gray-100 hover:text-gray-900 focus:outline-none focus:bg-gray-100 focus:text-gray-900",
-                  selectedValues.includes(option) &&
-                    "bg-primary-100 hover:bg-primary-200"
+                  "w-full block px-4 py-2 text-sm leading-5 text-left text-gray-700 hover:text-gray-900 focus:outline-none focus:text-gray-900",
+                  selectedValues.includes(option)
+                    ? "bg-primary-100 hover:bg-primary-200"
+                    : "hover:bg-gray-100 focus:bg-gray-100"
                 )}
                 onClick={() => {
                   setSelectedValues(
