+++ src/Components/Facility/Consultations/PrimaryParametersPlot.tsx
@@ -9,7 +9,8 @@ import Pagination from "../../Common/Pagination";
 import { PAGINATION_LIMIT } from "../../../Common/constants";
 import { formatDate } from "../../../Utils/utils";
 import CareIcon from "../../../CAREUI/icons/CareIcon";
-
+import { PainDiagrams } from "./PainDiagrams";
+import PageTitle from "../../Common/PageTitle";
 interface PrimaryParametersPlotProps {
   facilityId: string;
   patientId: string;
@@ -31,7 +32,6 @@ export const PrimaryParametersPlot = ({
           {
             page: currentPage,
             fields: [
-              "pain",
               "bp",
               "pulse",
               "temperature",
@@ -139,6 +139,7 @@ export const PrimaryParametersPlot = ({
       rhythmValues[key] = lst;
     }
   });
+
   return (
     <div>
       <div className="grid md:grid-cols-2 gap-4">
@@ -256,7 +257,10 @@ export const PrimaryParametersPlot = ({
           </div>
         </div>
       </div>
-
+      <div>
+        <PageTitle title="Pain Scale" hideBack={true} breadcrumbs={false} />
+        <PainDiagrams consultationId={consultationId} />
+      </div>
       {totalCount > PAGINATION_LIMIT && (
         <div className="mt-4 flex w-full justify-center">
           <Pagination
