+++ src/Components/Patient/SampleFilters.tsx
@@ -12,7 +12,15 @@ import { getAnyFacility } from "../../Redux/actions";
 import { useDispatch } from "react-redux";
 import { CircularProgress } from "@material-ui/core";
 import useMergeState from "../../Common/hooks/useMergeState";
-import FilterButtons from "../Common/FilterButtons";
+import FiltersSlideover from "../../CAREUI/interactive/FiltersSlideover";
+
+const clearFilterState = {
+  status: "",
+  result: "",
+  facility: "",
+  facility_ref: null,
+  sample_type: "",
+};
 
 export default function UserFilter(props: any) {
   const { filter, onChange, closeFilter } = props;
@@ -63,88 +71,82 @@ export default function UserFilter(props: any) {
   }, [dispatch]);
 
   return (
-    <div>
-      <div className="pb-10">
-        <FilterButtons
-          onClose={closeFilter}
-          onApply={applyFilter}
-          onClear={() => {
-            navigate("/sample");
-            setFilterState(filterState);
-            closeFilter();
-          }}
+    <FiltersSlideover
+      advancedFilter={props}
+      onApply={applyFilter}
+      onClear={() => {
+        navigate("/sample");
+        setFilterState(clearFilterState);
+        closeFilter();
+      }}
+    >
+      <div className="w-full flex-none">
+        <div className="text-sm font-semibold">Status</div>
+        <LegacySelectField
+          name="status"
+          variant="outlined"
+          margin="dense"
+          value={filterState.status || 0}
+          options={[
+            { id: "", text: "SELECT" },
+            ...SAMPLE_TEST_STATUS.map(({ id, text }) => {
+              return { id, text: text.replaceAll("_", " ") };
+            }),
+          ]}
+          onChange={handleChange}
+          errors=""
         />
       </div>
-      <div className="font-light text-md mt-2">Filter By:</div>
-      <div className="flex flex-wrap gap-2">
-        <div className="w-full flex-none">
-          <div className="text-sm font-semibold">Status</div>
-          <LegacySelectField
-            name="status"
-            variant="outlined"
-            margin="dense"
-            value={filterState.status || 0}
-            options={[
-              { id: "", text: "SELECT" },
-              ...SAMPLE_TEST_STATUS.map(({ id, text }) => {
-                return { id, text: text.replaceAll("_", " ") };
-              }),
-            ]}
-            onChange={handleChange}
-            errors=""
-          />
-        </div>
 
-        <div className="w-full flex-none">
-          <div className="text-sm font-semibold">Result</div>
-          <LegacySelectField
-            name="result"
-            variant="outlined"
-            margin="dense"
-            value={filterState.result || 0}
-            options={[{ id: "", text: "SELECT" }, ...SAMPLE_TEST_RESULT]}
-            onChange={handleChange}
-            errors=""
-          />
-        </div>
+      <div className="w-full flex-none">
+        <div className="text-sm font-semibold">Result</div>
+        <LegacySelectField
+          name="result"
+          variant="outlined"
+          margin="dense"
+          value={filterState.result || 0}
+          options={[{ id: "", text: "SELECT" }, ...SAMPLE_TEST_RESULT]}
+          onChange={handleChange}
+          errors=""
+        />
+      </div>
 
-        <div className="w-full flex-none">
-          <div className="text-sm font-semibold">Sample Test Type</div>
-          <LegacySelectField
-            name="sample_type"
-            variant="outlined"
-            margin="dense"
-            value={filterState.sample_type}
-            options={[{ id: "", text: "SELECT" }, ...SAMPLE_TYPE_CHOICES]}
-            onChange={handleChange}
-            errors=""
-          />
-        </div>
+      <div className="w-full flex-none">
+        <div className="text-sm font-semibold">Sample Test Type</div>
+        <LegacySelectField
+          name="sample_type"
+          variant="outlined"
+          margin="dense"
+          value={filterState.sample_type}
+          options={[{ id: "", text: "SELECT" }, ...SAMPLE_TYPE_CHOICES]}
+          onChange={handleChange}
+          errors=""
+        />
+      </div>
 
-        <div className="w-full flex-none">
-          <span className="text-sm font-semibold">Facility</span>
-          <div className="">
-            {isFacilityLoading ? (
-              <CircularProgress size={20} />
-            ) : (
-              <FacilitySelect
-                multiple={false}
-                name="facility"
-                selected={filterState.facility_ref}
-                showAll={true}
-                setSelected={(obj) =>
-                  setFilterState({
-                    facility: (obj as FacilityModel)?.id,
-                    facility_ref: obj,
-                  })
-                }
-                className="shifting-page-filter-dropdown"
-                errors={""}
-              />
-            )}
-          </div>
+      <div className="w-full flex-none">
+        <span className="text-sm font-semibold">Facility</span>
+        <div className="">
+          {isFacilityLoading ? (
+            <CircularProgress size={20} />
+          ) : (
+            <FacilitySelect
+              multiple={false}
+              name="facility"
+              selected={filterState.facility_ref}
+              showAll={true}
+              setSelected={(obj) =>
+                setFilterState({
+                  facility: (obj as FacilityModel)?.id,
+                  facility_ref: obj,
+                })
+              }
+              className="shifting-page-filter-dropdown"
+              errors={""}
+            />
+          )}
         </div>
       </div>
-    </div>
+    </FiltersSlideover>
   );
 }
