+++ src/Components/Users/UserFilter.tsx
@@ -4,14 +4,13 @@ import { getDistrict } from "../../Redux/actions";
 import { navigate } from "raviger";
 import DistrictSelect from "../Facility/FacilityFilter/DistrictSelect";
 import parsePhoneNumberFromString from "libphonenumber-js";
-import FilterButtons from "../Common/FilterButtons";
 import TextFormField from "../Form/FormFields/TextFormField";
 import SelectMenuV2 from "../Form/SelectMenuV2";
 import { FieldLabel } from "../Form/FormFields/FormField";
 import { USER_TYPE_OPTIONS } from "../../Common/constants";
-import CareIcon from "../../CAREUI/icons/CareIcon";
 import useMergeState from "../../Common/hooks/useMergeState";
 import PhoneNumberFormField from "../Form/FormFields/PhoneNumberFormField";
+import FiltersSlideover from "../../CAREUI/interactive/FiltersSlideover";
 
 const parsePhoneNumberForFilterParam = (phoneNumber: string) => {
   if (!phoneNumber) return "";
@@ -84,89 +83,74 @@ export default function UserFilter(props: any) {
     setFilterState({ ...filterState, [name]: value });
 
   return (
-    <div className="pb-10 -mt-4">
-      <FilterButtons
-        onClose={closeFilter}
-        onApply={applyFilter}
-        onClear={() => {
-          navigate("/users");
-          setFilterState(clearFilterState);
-          closeFilter();
-        }}
+    <FiltersSlideover
+      advancedFilter={props}
+      onApply={applyFilter}
+      onClear={() => {
+        navigate("/users");
+        setFilterState(clearFilterState);
+        closeFilter();
+      }}
+    >
+      <TextFormField
+        label="First Name"
+        labelClassName="text-sm"
+        name="first_name"
+        value={filterState.first_name}
+        onChange={handleChange}
+        errorClassName="hidden"
+      />
+      <TextFormField
+        label="Last Name"
+        labelClassName="text-sm"
+        name="last_name"
+        value={filterState.last_name}
+        onChange={handleChange}
+        errorClassName="hidden"
       />
 
-      <div className="pt-20 text-md my-6 flex items-center text-gray-700 gap-2">
-        <CareIcon className="care-l-filter text-lg" />
-        <p>Filter by</p>
+      <div className="w-full flex-none">
+        <FieldLabel>Role</FieldLabel>
+        <SelectMenuV2
+          id="role"
+          placeholder="Show all"
+          options={USER_TYPE_OPTIONS}
+          optionLabel={(o) => o.role + ((o.readOnly && " (Read Only)") || "")}
+          optionValue={(o) => o.id}
+          value={filterState.user_type}
+          onChange={(v) => setFilterState({ ...filterState, user_type: v })}
+        />
       </div>
 
-      <div className="flex flex-wrap gap-4">
-        <TextFormField
-          className="w-full"
-          label="First Name"
-          labelClassName="text-sm"
-          name="first_name"
-          value={filterState.first_name}
-          onChange={handleChange}
-          errorClassName="hidden"
+      <div className="w-full flex-none">
+        <FieldLabel>District</FieldLabel>
+        <DistrictSelect
+          multiple={false}
+          name="district"
+          selected={filterState.district_ref}
+          setSelected={setDistrict}
+          className="shifting-page-filter-dropdown"
+          errors={""}
         />
-        <TextFormField
-          className="w-full"
-          label="Last Name"
-          labelClassName="text-sm"
-          name="last_name"
-          value={filterState.last_name}
-          onChange={handleChange}
-          errorClassName="hidden"
-        />
-
-        <div className="w-full flex-none">
-          <FieldLabel className="text-sm">Role</FieldLabel>
-          <SelectMenuV2
-            id="role"
-            placeholder="Show all"
-            options={USER_TYPE_OPTIONS}
-            optionLabel={(o) => o.role + ((o.readOnly && " (Read Only)") || "")}
-            optionValue={(o) => o.id}
-            value={filterState.user_type}
-            onChange={(v) => setFilterState({ ...filterState, user_type: v })}
-          />
-        </div>
-
-        <div className="w-full flex-none">
-          <FieldLabel className="text-sm">District</FieldLabel>
-          <DistrictSelect
-            multiple={false}
-            name="district"
-            selected={filterState.district_ref}
-            setSelected={setDistrict}
-            className="shifting-page-filter-dropdown"
-            errors={""}
-          />
-        </div>
-
-        <div className="flex flex-wrap">
-          <div className="w-full flex-none">
-            <PhoneNumberFormField
-              label="Phone Number"
-              name="phone_number"
-              placeholder="Phone Number"
-              value={filterState.phone_number}
-              onChange={handleChange}
-            />
-          </div>
-
-          <div className="w-full flex-none -mt-2">
-            <PhoneNumberFormField
-              label="Whatsapp Number"
-              name="alt_phone_number"
-              placeholder="WhatsApp Phone Number"
-              value={filterState.alt_phone_number}
-              onChange={handleChange}
-            />
-          </div>
-        </div>
       </div>
-    </div>
+
+      <PhoneNumberFormField
+        label="Phone Number"
+        name="phone_number"
+        placeholder="Phone Number"
+        value={filterState.phone_number}
+        onChange={handleChange}
+        errorClassName="hidden"
+      />
+
+      <PhoneNumberFormField
+        label="Whatsapp Number"
+        name="alt_phone_number"
+        placeholder="WhatsApp Phone Number"
+        value={filterState.alt_phone_number}
+        onChange={handleChange}
+        errorClassName="hidden"
+      />
+    </FiltersSlideover>
   );
 }
