+++ src/service-worker.ts
@@ -10,31 +10,36 @@
 
 import { clientsClaim } from "workbox-core";
 import { ExpirationPlugin } from "workbox-expiration";
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
 import { precacheAndRoute } from "workbox-precaching";
 // import { createHandlerBoundToURL } from "workbox-precaching";
 import { registerRoute } from "workbox-routing";
 import { StaleWhileRevalidate } from "workbox-strategies";
-import { NetworkFirst } from "workbox-strategies";
 
 declare const self: ServiceWorkerGlobalScope;
 
+// eslint-disable-next-line no-restricted-globals
+const _ignored = self.__WB_MANIFEST.map((_) => {
+  return _;
+});
+
 clientsClaim();
 
-const toPrecache = self.__WB_MANIFEST.filter((file) => {
-  // Check if file is a string
-  if (typeof file === "string") {
-    return !file.includes("index.html");
-  } else {
-    return !file.url.includes("index.html");
-  }
-});
+// const toPrecache = self.__WB_MANIFEST.filter((file) => {
+//   // Check if file is a string
+//   if (typeof file === "string") {
+//     return !file.includes("index.html");
+//   } else {
+//     return !file.url.includes("index.html");
+//   }
+// });
 
 // Precache all of the assets generated by your build process.
 // Their URLs are injected into the manifest variable below.
 // This variable must be present somewhere in your service worker file,
 // even if you decide not to use precaching. See https://cra.link/PWA
 // precacheAndRoute(self.__WB_MANIFEST);
-precacheAndRoute(toPrecache);
+// precacheAndRoute(toPrecache);
 
 // Set up App Shell-style routing, so that all navigation requests
 // are fulfilled with your index.html shell. Learn more at
@@ -65,10 +70,10 @@ precacheAndRoute(toPrecache);
 //   createHandlerBoundToURL(import.meta.env.REACT_PUBLIC_URL + "/index.html")
 // );
 
-registerRoute(
-  ({ url }) => url.pathname.includes("index.html"),
-  new NetworkFirst()
-);
+// registerRoute(
+//   ({ url }) => url.pathname.includes("index.html"),
+//   new NetworkFirst()
+// );
 
 // An example runtime caching route for requests that aren't handled by the
 // precache, in this case same-origin .png requests like those from in public/
